RA-92: Resolved merge conflict

+++ api/src/test/java/org/cardanofoundation/rosetta/api/construction/service/PreprocessApiTest.java
@@ -52,7 +52,7 @@ class PreprocessApiTest extends IntegrationTest {
 
     ConstructionPreprocessResponse constructionPreprocessResponse = constructionApiService.constructionPreprocessService(
         preprocessRequest);
-    Map<String, Double> options = (Map<String, Double>) constructionPreprocessResponse.getOptions();
+    Map<String, Integer> options = (Map<String, Integer>) constructionPreprocessResponse.getOptions();
     assertEquals(expectedTtl, options.get("relative_ttl"));
     assertEquals(expectedTransactionSize, options.get("transaction_size"));
   }
